Split out raw ffi bindings into a separate crate
This commit splits out the raw bindings to Qiskit's C api into a
separate crate: qiskit-sys. Using a separate crate for the raw bindings
is a fairly common pattern for rust bindings to C libraries, it offers
a few advantages including:

 - it enables people to build against the raw C API if they don't want
   to use our crate
 - it isolates the build system complexity, like sourcing the C library
   and header, invoking bindgen, etc.
 - it facilitates better compilation caching since the raw bindings
   only change once per upstream Qiskit release so we don't need to
   regenerate and rebuild it as frequently

Additionally the build config is updated to not commit the raw bindings
into the repo and just let bindgen generate it at build time from the
header. This makes the wrapper and logic around the bindings simpler and
also reduces the size of the repo.