Add missing sys crate and commit missing sys crate
The previous 81e183fd1cb6fa24ff7d855a216bd858ad6090f3 we forgot to add
the new sys crate that was referred to in the PR. This commit adds the
missing module. It also runs `cargo fmt` on the repo and fixes the
reported clippy errors and warnings, except for the missing docs
warnings.