Add license file and apache headers to source files
This project is intended to be an open source library for using Qiskit
from Rust via the C API. However it was missing the license file and the
license headers in each source file. We were calling out the library as
apache licensed in the package metadata but not inidicating it anywhere
else in the source tree. This commit fixes this oversight and adds an
Apache 2.0 license file and appropriate file headers to each file.