Add basic CI configuration
This commit adds a basic CI configuration for code formatting,
lint checks, and running tests. This is just an initial configuration
we'll probably expand this in future work to enable multiple version tests. Especially after we define an MSRV for the library. But this keeps the configuration simple to start so we at least are validating everything before merging.